+++ WebStore/Services/InMemory/InMemoryProductData.cs
@@ -7,7 +7,7 @@ using WebStore.Domain;
 using WebStore.Domain.Entities;
 using WebStore.Services.Interfaces;
 
-namespace WebStore.Services
+namespace WebStore.Services.InMemory
 {
     public class InMemoryProductData : IProductData
     {
@@ -29,7 +29,7 @@ namespace WebStore.Services
             return query;
         }
 
-        
-        
+
+
     }
 }
