+++ WebStore/WebStore.csproj
@@ -20,4 +20,8 @@
     <ProjectReference Include="..\WebStore.Domain\WebStore.Domain.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Services\InSQL\" />
+  </ItemGroup>
+
 </Project>
